refactor: "extract class" type refactor to maintain responsiblity
